Create Trainer updater.py
The above code is for a Trainer class that can be used to train a neural network model. The Trainer class takes in a model, a train dataset, a test dataset, and a configuration object (of class TrainerConfig) as input and has the following functions:

__init__: Initializes the Trainer class and moves the model to the correct device (GPU if available, otherwise CPU).
save_checkpoint: Saves the model's state dictionary to the specified checkpoint path.
run_epoch: Runs a single epoch of training or evaluation, depending on the value of the split argument. This function loops over the data in the specified dataset, passes the input through the model to get the logits and loss, and updates the model's parameters if split is 'train'.
The TrainerConfig class is a simple configuration class that holds various hyperparameters and other settings for the training process. The default values for these parameters are specified in the class definition, but they can also be overridden by passing in keyword arguments to the constructor.

Note that the train function is not implemented in the above code, so it is not clear how the run_epoch function is being used to train the model. To complete the training loop, you will need to define a train function that calls run_epoch in a loop over the number of epochs specified in the TrainerConfig object, and that also handles things like setting up the optimizer, scheduling the learning rate, and possibly saving checkpoints.